Added Payments fixes
